CPU Sampler
This commit changes the Sampler interface to return map[string]int64.
This allows a sampler to return multiple values, which will allow
support for more complicated samplers, such as a CPU sampler which
returns multiple values. This also requires a rework of the handling
of deltas, which actually simplifies the code.

This commit also introduces a CPU sampler, which in a single config
item, handles all values for a given cpu. This could be overall or
per-CPU, and improves performance as well as the maintainability of
the configuration file.